server/requirements.txt
absl-py==2.0.0
amqp==5.1.1
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
awscli==1.29.68
beautifulsoup4==4.12.2
billiard==4.1.0
blinker==1.6.3
boto3==1.28.68
botocore==1.31.68
cachetools==5.3.1
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
configobj==5.0.8
constantly==23.10.4
contourpy==1.1.1
cryptography==41.0.5
cycler==0.12.1
daphne==4.0.0
deepface==0.0.79
dj-database-url==2.1.0
Django==4.2.6
django-celery-results==2.5.1
django-cors-headers==4.3.0
django-environ==0.11.2
djangorestframework==3.14.0
docutils==0.16
drf-yasg==1.21.7
filelock==3.12.4
fire==0.5.0
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.43.1
gast==0.5.4
gdown==4.7.1
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
gunicorn==21.2.0
h5py==3.10.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
keras==2.14.0
kiwisolver==1.4.5
kombu==5.3.2
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
ml-dtypes==0.2.0
msgpack==1.0.7
mtcnn==0.1.1
numpy==1.26.1
oauthlib==3.2.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
protobuf==4.24.4
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pyngrok==7.0.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rsa==4.7.2
s3transfer==0.7.0
service-identity==23.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
termcolor==2.3.0
tqdm==4.66.1
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.8
websockets==12.0
Werkzeug==3.0.0
whitenoise==6.6.0
wrapt==1.14.1
zope.interface==6.1
